refactor(test): extract ADR-089 E2E helpers and fix PR test matrix names
Move shared Hub wiring to tests/integration/adr089_helpers.py so OMP and
generic E2E tests no longer import from each other.
