+++ tests/integration/test_adr089_error_resolution_e2e.py
@@ -6,9 +6,7 @@ and resolve_user_error() wiring. No NATS / no live CLI.
 
 from __future__ import annotations
 
-import asyncio
 from collections.abc import AsyncIterator
-from pathlib import Path
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
@@ -16,77 +14,23 @@ import pytest
 from factory.agents.simple_agent import SimpleAgent
 from factory.core.agent import Agent
 from factory.core.agent.agent_config import ModelConfig
-from factory.core.hub import Hub
-from factory.core.lifecycle.circuit_breaker import CircuitBreaker, CircuitRegistry
 from factory.core.messaging.events import ResultLlmEvent
 from factory.core.messaging.message import Platform
-from factory.core.messaging.messages import MessageManager
-from factory.core.messaging.render_events import (
-    RunErrorRenderEvent,
-    TextDeltaRenderEvent,
-)
 from factory.llm.base import LlmResult
 from roxabi_contracts.errors import WorkerError
-from tests.core.conftest import push_to_hub
-from tests.integration.test_e2e_telegram_to_agent import (
-    _make_telegram_message,
-    _RecordingAdapter,
+from tests.integration.adr089_helpers import (
+    DrainingStreamingAdapter,
+    hub_with_agent,
+    message_manager,
+    run_hub_until_processed,
 )
-
-_MESSAGES = (
-    Path(__file__).resolve().parents[2] / "src" / "factory" / "data" / "messages.toml"
-)
-
-
-def _message_manager() -> MessageManager:
-    return MessageManager(_MESSAGES, language="en")
-
-
-class _DrainingStreamingAdapter(_RecordingAdapter):
-    """Records outbound payloads and drains render-event streams."""
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.run_error_messages: list[str] = []
-        self.streamed_deltas: list[str] = []
-
-    async def send_streaming(self, original_msg, events, outbound=None) -> None:
-        del original_msg
-        async for event in events:
-            if isinstance(event, RunErrorRenderEvent):
-                self.run_error_messages.append(event.message)
-            elif isinstance(event, TextDeltaRenderEvent):
-                self.streamed_deltas.append(event.delta)
-        if outbound is not None:
-            self.streamed.append(outbound)
-
-
-def _hub_with_agent(agent: SimpleAgent) -> Hub:
-    mm = _message_manager()
-    registry = CircuitRegistry()
-    registry.register(CircuitBreaker("claude-cli"))
-    hub = Hub(circuit_registry=registry, msg_manager=mm)
-    hub.register_agent(agent)
-    adapter = _DrainingStreamingAdapter()
-    hub.register_adapter(Platform.TELEGRAM, "main", adapter)
-    hub.register_binding(
-        Platform.TELEGRAM, "main", "*", agent.config.name, "telegram:main:*"
-    )
-    return hub
-
-
-async def _run_hub_until_processed(hub: Hub) -> None:
-    await push_to_hub(hub, _make_telegram_message("hi"))
-    try:
-        await asyncio.wait_for(hub.run(), timeout=2.0)
-    except asyncio.TimeoutError:
-        pass
+from tests.integration.test_e2e_telegram_to_agent import _RecordingAdapter
 
 
 @pytest.mark.smoke
 class TestAdr089BlockingErrorE2E:
     async def test_rate_limit_template_reaches_telegram_adapter(self) -> None:
-        mm = _message_manager()
+        mm = message_manager()
         provider = MagicMock()
         provider.complete = AsyncMock(
             return_value=LlmResult(
@@ -109,11 +53,11 @@ class TestAdr089BlockingErrorE2E:
             provider,
             msg_manager=mm,
         )
-        hub = _hub_with_agent(agent)
+        hub = hub_with_agent(agent)
         adapter = hub.adapter_registry[(Platform.TELEGRAM, "main")]
         assert isinstance(adapter, _RecordingAdapter)
 
-        await _run_hub_until_processed(hub)
+        await run_hub_until_processed(hub)
 
         assert len(adapter.sent) == 1
         assert adapter.sent[0].to_text() == mm.get("rate_limit")
@@ -140,13 +84,13 @@ class TestAdr089BlockingErrorE2E:
                 llm_config=ModelConfig(streaming=False),
             ),
             provider,
-            msg_manager=_message_manager(),
+            msg_manager=message_manager(),
         )
-        hub = _hub_with_agent(agent)
+        hub = hub_with_agent(agent)
         adapter = hub.adapter_registry[(Platform.TELEGRAM, "main")]
         assert isinstance(adapter, _RecordingAdapter)
 
-        await _run_hub_until_processed(hub)
+        await run_hub_until_processed(hub)
 
         assert len(adapter.sent) == 1
         assert "weekly limit" in adapter.sent[0].to_text()
@@ -165,13 +109,13 @@ class TestAdr089BlockingErrorE2E:
                 llm_config=ModelConfig(streaming=False),
             ),
             provider,
-            msg_manager=_message_manager(),
+            msg_manager=message_manager(),
         )
-        hub = _hub_with_agent(agent)
+        hub = hub_with_agent(agent)
         adapter = hub.adapter_registry[(Platform.TELEGRAM, "main")]
         assert isinstance(adapter, _RecordingAdapter)
 
-        await _run_hub_until_processed(hub)
+        await run_hub_until_processed(hub)
 
         assert len(adapter.sent) == 1
         text = adapter.sent[0].to_text()
@@ -183,7 +127,7 @@ class TestAdr089BlockingErrorE2E:
 @pytest.mark.smoke
 class TestAdr089StreamingSoftErrorE2E:
     async def test_rate_limit_soft_error_reaches_streaming_adapter(self) -> None:
-        mm = _message_manager()
+        mm = message_manager()
         expected = mm.get("rate_limit")
 
         async def _error_stream() -> AsyncIterator[ResultLlmEvent]:
@@ -212,11 +156,11 @@ class TestAdr089StreamingSoftErrorE2E:
             provider,
             msg_manager=mm,
         )
-        hub = _hub_with_agent(agent)
+        hub = hub_with_agent(agent)
         adapter = hub.adapter_registry[(Platform.TELEGRAM, "main")]
-        assert isinstance(adapter, _DrainingStreamingAdapter)
+        assert isinstance(adapter, DrainingStreamingAdapter)
 
-        await _run_hub_until_processed(hub)
+        await run_hub_until_processed(hub)
 
         assert len(adapter.streamed) == 1
         assert adapter.run_error_messages == [expected]
